feat: fix 100% verification circle color and add Verification filter
- Cast round() to (int) so strict === 100 comparison works for green color
- Add Verification dropdown filter (Fully Verified / Partially / Not Verified)
- Filter computes verification % post-fetch using LengthAwarePaginator

+++ resources/views/backend/pages/pending-approvals/index.blade.php
@@ -92,7 +92,7 @@
                                         if (in_array('cf_' . $scf->id, $verifiedFields, true)) $vDone++;
                                     }
                                 }
-                                $vPct = $vTotal > 0 ? round(($vDone / $vTotal) * 100) : 0;
+                                $vPct = $vTotal > 0 ? (int) round(($vDone / $vTotal) * 100) : 0;
                             @endphp
                             <tr class="hover:bg-gray-50 dark:hover:bg-gray-700/50">
                                 <td class="px-4 py-3 whitespace-nowrap">
